made final changes to the app
after creating the basic app now all the final touches were added to the app. some of the functionalities that were not working before are corrected. the look of the app remains the same. some new text and features are added to the app. now in reminders tab the user can set an alarm. in events and timetables pages new text is added to guide the user about the add button below. the look of how a day and date/time is also changed, now it looks more neat. datebase was linked with reminders and events page. the delete button also now deletes anything in real time. the dynamic calendar that was added in the previous commit also works fine and changes the day and date dynamically.